feat: Panel애니메이션 추가
